fix(workout-store): switch to long values and fix flow to restart workout

+++ shared/src/commonMain/kotlin/com/readysetmove/personalworkouts/workout/Set.kt
@@ -1,3 +1,3 @@
 package com.readysetmove.personalworkouts.workout
 
-data class Set(val tractionGoal: Float, val duration: Float, val restTime: Float)
+data class Set(val tractionGoal: Long, val duration: Long, val restTime: Long)
